03-deep_learning/01-tensorflow_examples/04-CV/04-ocr/OCR_repo/requirements.txt
tensorflow_gpu==1.6.0
numpy==1.13.1
opencv_python==3.2.0.7
matplotlib==2.0.2
keras==2.1.5
captcha==0.2.4
Pillow==5.1.0
pandas==0.23.0
tqdm==4.24.0



